% Auto-generated: do not edit by hand
\name{standardButton}

\alias{standardButton}

\title{StandardButton component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
standardButton(children=NULL, id=NULL, onClick=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{onClick}{Character. On click callback (must be a string, as Dash components user defined props must be serializable)}
}
